Added a PassengerController class and copied code from PassengerApplication into it
