+++ src/main/java/ie/atu/PassengerJPA/PassengerJpaApplication.java
@@ -11,28 +11,10 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.List;
 
 @RestController
-@SpringBootApplication
-@RequestMapping(path="api/passenger")
+
 public class PassengerJpaApplication {
-
-	PassengerService myService;
-
-	public PassengerJpaApplication(PassengerService myService) {
-		this.myService = myService;
-	}
-
 	public static void main(String[] args) {
 		SpringApplication.run(PassengerJpaApplication.class, args);
-	}
-
-	@GetMapping
-	public List<Passenger> getPassengers(){
-		return myService.getPassengers();
-	}
-
-	@GetMapping("/{passengerID}")
-	public Passenger getPassenger(@PathVariable String passengerID){
-		return myService.getPassenger(passengerID);
 	}
 }
